ai_model/requirements.txt
Flask
pdfplumber
python-docx
opencv-python
pytesseract
Pillow
numpy
mammoth
PyMuPDF
pdfkit
html2docx
sumy
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
reportlab
gunicorn